Remove message logging
Please note that this commit is strictly about logging what happens with
a message (message #1 received, message #3 closed, message #2 retried,
etc...)

During production testing I've found that I can't use this logging
properly. I rather implemented the required logging manually. So the
idea is that the users will implement the logging they need and I
don't try to come up with a universal logging solution. It would be
either too little or too much.

What remained is the exception tracking which is moved to the ITracer
interface. I hope that interface will be enough for all
instrumentation in the future.